refactor: extract tour-target retry polling into useTourTargetRetry hook
The TutorialProvider monolith had a ~40-line useEffect that polled the
DOM every 250ms for a missing step target to appear, with timeout and
ref-based cancellation. Moved it into a custom hook alongside the
existing useTutorialAutoScroll.

The constants (MAX_TARGET_RETRIES, RETRY_INTERVAL_MS) and the logger
context move with the effect. TutorialProvider now reads top-to-bottom
without the polling implementation interleaved with tour state logic.

Pure structural extraction — behavior identical. 9/9 TutorialProvider
tests pass unchanged.
